requirements.txt
langchain==0.3.25
langchain-community==0.3.24
langchain-anthropic==0.3.15
langchain-chroma==0.2.4
langchain-huggingface>=0.1.0
fastapi==0.115.12
uvicorn[standard]==0.34.3
python-multipart==0.0.20
chromadb>=1.0.0
sentence-transformers>=2.2.0
pypdf==5.4.0
python-docx==1.1.2
python-dotenv==1.1.0
pydantic-settings==2.9.1
Pillow>=10.0.0
